feat: move auth pages to 'pages' dir, apply 'withAuth' HOC
